requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
click==7.0
importlib-metadata==0.18  # via pluggy, pytest
more-itertools==7.0.0     # via pytest
packaging==19.0           # via pytest
pillow==6.0.0
pluggy==0.12.0            # via pytest
py==1.8.0                 # via pytest
pyparsing==2.4.0          # via packaging
pytest==4.6.3
six==1.12.0               # via packaging, pytest
wcwidth==0.1.7            # via pytest
zipp==0.5.1               # via importlib-metadata
